# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Full Sales Web System API
  version: 1.0.0
  description: Driver/seller control system — full REST surface (see docs/API-CONTRACT.md)
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://127.0.0.1:8080/v1
    description: Local development
paths:
...
  /portal/orders/{id}:
    get:
      summary: Get portal order
      operationId: getPortalOrder
      tags: [Portal]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order found
    put:
      summary: Update draft order
      operationId: updatePortalOrder
      tags: [Portal]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Draft updated
    delete:
      summary: Cancel draft order
      operationId: cancelPortalOrder
      tags: [Portal]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '204':
          description: Draft cancelled
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...